+++ pyevolver/tests/evolution_test.py
@@ -339,11 +339,41 @@ def test_v002():
     evo.run()
     evo.timing.report()
 
+def test_v003():
+
+    folder_path = './data/tmp1'
+    utils.make_dir_if_not_exists(folder_path)
+
+    evo = Evolution(
+        random_seed=np.random.randint(100000),
+        population_size=4,
+        genotype_size=2,
+        evaluation_function=lambda pop, seeds: np.sum(pop, axis=1), #np.arange(1,0,-1/len(pop)),
+        performance_objective='ABS_MAX', #MAX MIN ZERO ABS_MAX
+        fitness_normalization_mode='NONE',
+        selection_mode='UNIFORM', # UNIFORM, SUS, RWS
+        reproduce_from_elite=True,
+        reproduction_mode='GENETIC_ALGORITHM',  # 'GENETIC_ALGORITHM' 'HILL_CLIMBING'
+        mutation_variance=0.1,
+        elitist_fraction=0.5,
+        mating_fraction=0.5,
+        crossover_probability=0.5,
+        crossover_mode='UNIFORM',
+        max_generation=100,
+        termination_function=None,
+        checkpoint_interval=1,
+        folder_path=folder_path,
+        timeit=True
+    )
+    evo.run()
+    evo.timing.report()
+
 if __name__ == "__main__":
     # assertion_tests()
     # rws_test()
     # time_evolution()
     # test_continuation()
-    test_v002()
+    # test_v002()
+    test_v003()